refactor: move compose mail toolbar into a new component
