+++ frontend/components.d.ts
@@ -23,6 +23,7 @@ declare module 'vue' {
     AutocompleteControl: typeof import('./src/components/Controls/AutocompleteControl.vue')['default']
     ChangePasswordModal: typeof import('./src/components/Modals/ChangePasswordModal.vue')['default']
     ComposeMailEditor: typeof import('./src/components/ComposeMailEditor.vue')['default']
+    ComposeMailToolbar: typeof import('./src/components/ComposeMailToolbar.vue')['default']
     CopyCode: typeof import('./src/components/CopyCode.vue')['default']
     CopyControl: typeof import('./src/components/Controls/CopyControl.vue')['default']
     DashboardLayout: typeof import('./src/components/DashboardLayout.vue')['default']
