+++ frontend/src/components/ComposeMailEditor.vue
@@ -12,7 +12,7 @@
 	>
 		<template #top>
 			<div class="flex flex-col gap-2.5 border-b pb-2.5">
-				<div v-if="!mailDetails?.type" class="flex justify-between gap-2">
+				<div v-if="!mailDetails?.type || isMobile" class="flex justify-between gap-2">
 					<div class="flex items-center gap-2">
 						<span class="text-ink-gray-4 text-xs">{{ __('From') }}:</span>
 						<AutocompleteControl
@@ -99,7 +99,7 @@
 						</div>
 					</div>
 				</div>
-				<div v-if="!mailDetails?.type" class="flex items-center gap-2">
+				<div v-if="!mailDetails?.type || isMobile" class="flex items-center gap-2">
 					<span class="text-ink-gray-4 text-xs">{{ __('Subject') }}:</span>
 					<input
 						v-model="mail.subject"
@@ -152,81 +152,15 @@
 			</div>
 		</template>
 		<template #bottom>
-			<FileUploader
-				:class="{
-					'fixed left-0 right-0 z-20 px-3 transition-all': isMobile,
-				}"
-				:style="{ bottom: `${toolbarBottom}px` }"
-				:upload-args="{ private: true, folder: 'Home/Frappe Mail' }"
-				@success="(file) => mail.attachments.push({ ...file, disposition: 'attachment' })"
-			>
-				<template #default="{ file, progress, uploading, openFileSelector, error }">
-					<div
-						v-if="uploading"
-						class="bg-surface-gray-2 text-ink-gray-6 mb-2 rounded p-2.5 text-sm"
-					>
-						<div class="mb-1.5 flex items-center">
-							<span class="mr-1 font-medium"> {{ file.name }} </span>
-							<span class="font-extralight"> ({{ formatBytes(file.size) }}) </span>
-						</div>
-						<Progress :value="progress" />
-					</div>
-
-					<ErrorMessage :message="error" class="mb-2.5" />
-
-					<div
-						class="flex flex-wrap justify-between gap-2 overflow-hidden border-t pt-2.5"
-						:class="{ 'pb-2.5': isMobile }"
-					>
-						<!-- Text editor buttons -->
-						<div class="flex items-center gap-1 overflow-x-auto">
-							<TextEditorFixedMenu
-								:buttons="textEditorButtons"
-								class="!bg-inherit"
-							/>
-							<EmojiPicker
-								v-if="!isMobile"
-								v-slot="{ togglePopover }"
-								v-model="emoji"
-								@update:model-value="() => appendEmoji()"
-							>
-								<Button variant="ghost" @click="togglePopover()">
-									<template #icon>
-										<Laugh class="h-4 w-4" />
-									</template>
-								</Button>
-							</EmojiPicker>
-							<Button variant="ghost" @click="openFileSelector()">
-								<template #icon>
-									<Paperclip class="h-4" />
-								</template>
-							</Button>
-						</div>
-
-						<!-- Send & Discard -->
-						<div v-if="!isMobile" class="ml-auto flex items-center space-x-2">
-							<span v-if="isSavingDraft" class="text-ink-gray-5 text-base italic">
-								{{ __('Saving Draft...') }}
-							</span>
-							<template v-else>
-								<Button
-									:label="__('Discard')"
-									:icon-left="Trash2"
-									:disabled="isLoading"
-									@click="discardMail"
-								/>
-								<Button
-									variant="solid"
-									:label="__('Send')"
-									:icon-left="SendHorizontal"
-									:disabled="isRecipientsEmpty || isLoading"
-									@click="sendMail"
-								/>
-							</template>
-						</div>
-					</div>
-				</template>
-			</FileUploader>
+			<ComposeMailToolbar
+				:is-saving-draft="isSavingDraft"
+				:is-loading="isLoading"
+				:is-recipients-empty="isRecipientsEmpty"
+				@add-attachment="(file) => mail.attachments.push(file)"
+				@append-emoji="(emoji: string) => appendEmoji(emoji)"
+				@discard-mail="discardMail"
+				@send-mail="sendMail"
+			/>
 		</template>
 	</TextEditor>
 </template>
@@ -244,35 +178,16 @@ import {
 } from 'vue'
 import { EditorContent } from '@tiptap/vue-3'
 import { watchDebounced } from '@vueuse/core'
-import {
-	ExternalLink,
-	Forward,
-	Laugh,
-	Paperclip,
-	Reply,
-	ReplyAll,
-	SendHorizontal,
-	Trash2,
-} from 'lucide-vue-next'
-import {
-	Button,
-	Dropdown,
-	ErrorMessage,
-	FeatherIcon,
-	FileUploader,
-	Progress,
-	TextEditor,
-	TextEditorFixedMenu,
-	createResource,
-} from 'frappe-ui'
+import { ExternalLink, Forward, Reply, ReplyAll } from 'lucide-vue-next'
+import { Button, Dropdown, FeatherIcon, TextEditor, createResource } from 'frappe-ui'
 import { ImageExtension } from 'frappe-ui/src/components/TextEditor/extensions/image'
 import { useFileUpload } from 'frappe-ui/src/utils/useFileUpload'
 
-import { formatBytes, raiseToast, textEditorButtons, validateEmail } from '@/utils'
+import { formatBytes, raiseToast, validateEmail } from '@/utils'
 import { useScreenSize } from '@/utils/composables'
+import ComposeMailToolbar from '@/components/ComposeMailToolbar.vue'
 import AutocompleteControl from '@/components/Controls/AutocompleteControl.vue'
 import MultiselectInputControl from '@/components/Controls/MultiselectInputControl.vue'
-import EmojiPicker from '@/components/EmojiPicker.vue'
 
 import type { Attachment, ComposeMailData, File as FileDoc, UserResource } from '@/types'
 
@@ -286,6 +201,8 @@ const { mailDetails, isInThread = false } = defineProps<{
 const emit = defineEmits(['reloadMails', 'discardMail', 'reply', 'replyAll', 'forward', 'popOut'])
 
 // Editor
+
+const { isMobile } = useScreenSize()
 
 const textEditor = useTemplateRef('textEditor')
 const toInput = useTemplateRef('toInput')
@@ -304,11 +221,9 @@ const toggleBCC = () => {
 	if (bcc.value) nextTick(() => bccInput.value?.setFocus())
 }
 
-const emoji = ref()
-const appendEmoji = () => {
-	textEditor.value.editor.commands.insertContent(emoji.value)
+const appendEmoji = (emoji: string) => {
+	textEditor.value.editor.commands.insertContent(emoji)
 	textEditor.value.editor.commands.focus()
-	emoji.value = ''
 }
 
 // Setup & hooks
@@ -342,7 +257,7 @@ onMounted(() => {
 
 onUnmounted(() => {
 	saveDraft()
-	emit('discardMail')
+	// emit('discardMail')
 })
 
 watchDebounced(mail, () => saveDraft(), { debounce: 2000 })
@@ -445,34 +360,6 @@ const localDraftActions = computed(() => [
 		],
 	},
 ])
-
-// Make toolbar hover over keyboard on mobile
-
-const { isMobile } = useScreenSize()
-
-const toolbarBottom = ref(0)
-
-const updatePosition = () => {
-	if (!window.visualViewport) return
-	const offset =
-		window.innerHeight - window.visualViewport.height - window.visualViewport.offsetTop
-	toolbarBottom.value = offset > 0 ? offset : 0
-}
-
-onMounted(() => {
-	if (!(isMobile.value && window.visualViewport)) return
-
-	window.visualViewport.addEventListener('resize', updatePosition)
-	window.visualViewport.addEventListener('scroll', updatePosition)
-
-	updatePosition()
-
-	onUnmounted(() => {
-		if (!window.visualViewport) return
-		window.visualViewport.removeEventListener('resize', updatePosition)
-		window.visualViewport.removeEventListener('scroll', updatePosition)
-	})
-})
 
 // Mail content
